pyproject.toml
[project]
name = "rag-pdf"
version = "0.1.0"
description = "Socle RAG pour plans AutoCAD PDF avec citations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mistralai>=1.0,<2",
    "PyMuPDF>=1.24,<2",
    "python-dotenv>=1.0,<2",
    "streamlit>=1.44,<2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
